Move logic to geometric
