+++ geograypher/utils/geometric.py
@@ -2,6 +2,7 @@ import logging
 import typing
 
 import numpy as np
+import pyvista as pv
 import shapely
 from shapely import simplify, unary_union
 from tqdm import tqdm
@@ -138,3 +139,116 @@ def projection_onto_plane(v1, e1, e2):
     out_of_plane_projection = orthogonal_projection(plane_normal, v1)
     in_plane_projection = v1 - out_of_plane_projection
     return in_plane_projection
+
+
+def clip_line_segments(
+    boundaries: typing.Tuple[pv.PolyData, pv.PolyData],
+    origins: np.ndarray,
+    directions: np.ndarray,
+    image_indices: typing.Union[typing.List[int], np.ndarray],
+    ray_limit: typing.Optional[float] = None,
+) -> typing.Tuple[np.ndarray, np.ndarray, np.ndarray, np.ndarray]:
+    """
+    Clips line segments between two boundary surfaces using ray tracing,
+    keeping only the segments that intersect both surfaces.
+
+    Args:
+        boundaries (Tuple[pv.PolyData, pv.PolyData]):
+            A tuple containing two PyVista PolyData surfaces representing the boundaries.
+        origins (np.ndarray):
+            Array of ray origin points (shape: [N, 3]).
+        directions (np.ndarray):
+            Array of ray direction vectors (shape: [N, 3]).
+        image_indices (List[int]):
+            List of indices associated with each ray, used for identification.
+            Tracks which image each ray corresponded to.
+        ray_limit (Optional[float], optional):
+            If provided, segments longer than this value are excluded. Note that the
+            filtered distance is from the **original ray start to the second boundary**.
+            This is to mimic measuring from a camera (hypothetical ray source) to
+            the ground (assuming the boundaries are given as [ceiling, floor]).
+
+    Returns:
+        Tuple[np.ndarray, np.ndarray, np.ndarray, np.ndarray]:
+            - (N, 3) array of start points of clipped segments. This will be the point
+                that intersected boundaries[0]
+            - (N, 3) array of end points of clipped segments. This will be the point
+                that intersected boundaries[1]
+            - (N, 3) array of direction vectors for the clipped segments.
+            - (N,) array of image indices corresponding to the clipped segments.
+
+    Raises:
+        ValueError if various input requirements are not met.
+    """
+
+    # Input checking
+    if len(boundaries) != 2:
+        raise ValueError(f"2 boundaries required, not {len(boundaries)}")
+    if any([not isinstance(b, pv.PolyData) for b in boundaries]):
+        raise ValueError(f"pv.PolyData required, found {[type(b) for b in boundaries]}")
+    if origins.shape != directions.shape:
+        raise ValueError(
+            f"origins and directions mismatched ({origins.shape} != {directions.shape})"
+        )
+    if origins.shape[1] != 3:
+        raise ValueError(f"(N, 3) input arrays required, found {origins.shape}")
+    if len(origins) != len(image_indices):
+        raise ValueError(
+            f"origins and image indices mismatched ({len(origins)} !="
+            f" {len(image_indices)})"
+        )
+
+    # Handle the empty case
+    if len(origins) == 0:
+        return (
+            origins.copy(),
+            origins.copy(),
+            directions.copy(),
+            np.array(image_indices),
+        )
+
+    # Calculate ray/boundary intersections
+    pts0, idx0, _ = boundaries[0].multi_ray_trace(
+        origins=origins,
+        directions=directions,
+        first_point=True,
+        retry=True,
+    )
+    pts1, idx1, _ = boundaries[1].multi_ray_trace(
+        origins=origins,
+        directions=directions,
+        first_point=True,
+        retry=True,
+    )
+
+    # Find the ray indices that were found to intersect with both surfaces
+    matched_set = list(set(idx0).intersection(set(idx1)))
+
+    # Iterate through the rays with intersections and update the starts and ends
+    new_starts = []
+    new_ends = []
+    new_directions = []
+    new_indices = []
+
+    for ray_index in matched_set:
+        pt0 = pts0[np.where(idx0 == ray_index)[0][0]]
+        pt1 = pts1[np.where(idx1 == ray_index)[0][0]]
+
+        # Limit the length of the origin → boundary[1] vector if requested
+        if ray_limit is not None:
+            vector = origins[ray_index] - pt1
+            if np.linalg.norm(vector) > ray_limit:
+                continue
+
+        # Keep track of the outputs
+        new_starts.append(pt0)
+        new_ends.append(pt1)
+        new_directions.append((pt1 - pt0) / np.linalg.norm(pt1 - pt0))
+        new_indices.append(image_indices[ray_index])
+
+    return (
+        np.array(new_starts),
+        np.array(new_ends),
+        np.array(new_directions),
+        np.array(new_indices),
+    )
